+++ src/lib.rs
@@ -30,7 +30,9 @@
 #![feature(generator_trait)]
 #![feature(generators)]
 #![feature(never_type)]
-#![warn(missing_docs)]
+#![feature(or_patterns)]
+#![feature(unsafe_block_in_unsafe_fn)]
+#![warn(missing_docs, unsafe_op_in_unsafe_fn)]
 #![warn(clippy::pedantic)]
 #![allow(
     clippy::cast_possible_truncation,